feat: bet state and record (#65)
* feat: added betting system

Added a betting system to the game. User starts with $100. Can't change bet when game has started. Handles negative numbers and numbers larger than whats in the bankroll. Not thoroughly tested for functionality. Styling was added, but is very much subject to change if needed.

closes #32

* docs: documented all betting-related code

* feat: updated buttons to work at proper times

Hit and Stand are no longer active when opening the browser

* delete empty changes in systemview

* style: delete space in line

* style: remove empty change

* refactor: organized code

Moved initialization of bankroll and updateBankrollDisplay to BettingView

* fix: fixed duplicate method names

* Prettified Code!

* fix: updaded based on code review

---------

Co-authored-by: thomasrocha0 <thomaslrocha05@gmail.com>
Co-authored-by: Sentixxx <mouxinc@gmail.com>
Co-authored-by: Sentixxx <Sentixxx@users.noreply.github.com>